src/requirements.txt
coverage==7.9.1
django==5.2.11
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
flake8==7.2.0
gunicorn==23.0.0
isort==6.0.1
psycopg2-binary==2.9.10
pytest==7.4.4
python-decouple==3.8
requests==2.32.4
pytest-django==4.11.1
pytest-cov==6.2.1
pytest-factoryboy==2.7.0
boto3==1.38.36
django-storages==1.14.6
django-q2==1.8.0
black==25.1.0
django-debug-toolbar==5.2.0
git+https://github.com/akaytatsu/django-factory-generator.git
git+https://github.com/vert-capital/django_kafka.git
freezegun==1.5.2
requests-mock==1.10.0
pip-upgrader
django-cors-headers==4.7.0
uvicorn==0.34.3
django-fsm==3.0.0
pytest-watch==4.2.0
pillow==12.1.1
django-cas-ng==4.2.1
pytz==2024.2
